Fix warnings (#1792)
* [NL] DOF: explicit cast of size for small vector.

* [NL] Use double pair of braces in initialization.

* [PL] SD: Remove unused function parameter.

* [MeL] Avoid implicit conversion from vtkIdType.

* [PL] LIE: Use int for variable and component ids.

* [T] GML: Avoid implicit integer conversion.

+++ NumLib/DOF/LocalToGlobalIndexMap.cpp
@@ -209,7 +209,7 @@ LocalToGlobalIndexMap::LocalToGlobalIndexMap(
             "There are %d mesh subsets and %d components.",
             mesh_subsets.size(), global_component_ids.size());
 
-    for (int i = 0; i < global_component_ids.size(); ++i)
+    for (int i = 0; i < static_cast<int>(global_component_ids.size()); ++i)
     {
         auto const& mss = _mesh_subsets[i];
